[win32][squashme] Drop boost asio.hpp include
asio.hpp has to be included before Winsock2.h (or Windows.h).
For that reason, we've just included it in the compatibility headers.
The issue with that is that in some cases, those are included
within "extern C" wrappers.

For this reason, we'll have to handle asio.hpp usage on a case by
case basis.