[win32] Fix erasure_code dependency
erasure_code is using libcommon but isn't linked against it, while
libcommon isn't directly using erasure_code directly, yet it's
depending on it.

When building Windows dlls, all symbols have to be located and
circular dependencies avoided. For this reason, we're going to
update the dependency chains involving erasure_code.