[win32] Initialize WSA and TLS
This patch provides a constructor that will be called before
main(). It's initializing the Windows Socket API.

At the same time, we're preparing a TLS slot that can be used for C++
destructors as a workaround for a Mingw bug [1][2][3], which would
otherwise result in crashes as the according TLS memory would be
cleaned up before the destructors are invoked.

[1] https://sourceforge.net/p/mingw-w64/bugs/727/
[2] https://sourceforge.net/p/mingw-w64/bugs/527/
[3] https://sourceforge.net/p/mingw-w64/bugs/445/

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>