common: avoid CLOCK_*_COARSE warnings on win32
The CLOCK_*_COARSE clocks are Linux specific and are
expected to be faster. When not available, CLOCK_REALTIME
and CLOCK_MONOTONIC will be used as fallback.

Ceph currently assumes that those will be slow, excessively
emitting the following warning:
"Falling back to CLOCK_REALTIME, may be slow."

Mingw uses QueryPerformanceCounter behind the scenes, which is
already one of the fastest clocks available on Windows (as opposed
to the *Precise* alternatives), depending on the hardware clock
available (e.g. TSC, HPET, etc). For this reason, we're going
to define CLOCK_REALTIME_COARSE as CLOCK_REALTIME, same for
the monotonic clock.

Signed-off-by: Alin Gabriel Serdean <aserdean@cloudbasesolutions.com>