[win32] port unit tests
We're porting to Windows some of the unit tests, mostly related to rados
and rbd.

Overall, this is about using compat.h functions and types as well as avoiding
unsupported functionality.

The systest framework will have to be limitted to threads instead of separate
processes until we cover the process spawning and communication.