[win32] Fix subprocess helper
The subprocess helper isn't retrieving the pid after spawning
subprocesses, for which reason "join" will error out.

This change will retrieve the pid using the process handle. While
at it, we'll set an error message when the return code is non-zero,
which is expected by the unit tests.

We're also updating the mechanism that's supposed to enforce timeouts.

Argument parsing also needs to be fixed.