cephfs,win32: Add ceph-dokan
In order to expose ceph filesystems to Windows hosts, we propose
including ceph-dokan[1] in the Ceph tree, while updating it to
work with the latest CephFS and Dokany APIs.

Dokany is a well maintained project (fork of the original Dokan
project), allowing filesystems to be implemented in userspace,
even providing a Fuse compatibility layer.

One reason for not using the Fuse compatibility layer is that it's
only covering the high level API while Ceph is using the low level
Ceph API, which among other things is inode centric.

Changes made by this patch compared to the upstream ceph-dokan:
* support latest stable Dokany API. The upstream version relies on
  the legacy unmaintained Dokan API
* minor changes to support latest cephfs API
* drop duplicated ceph code, no longer needed if we're to include it
  in tree. This makes it much easier to maintain.
* fixed unicode handling
* various cleanups

[1] https://github.com/ketor/ceph-dokan