win32: Fix compiler warnings
This change will fix a few compiler warnings, most of them being
type mismatches that affect Windows builds.

Worth mentioning that "long" uses 32b on x64 Windows platforms,
which is a constant source of type mismatches.

There's a warning saying that the "select" event handler shouldn't
be used in production. Considering that this is the only available
mechanism on Windows for the time being, we're going to disable
this warning.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>