win32, cephfs: configurable dokan timeout
ceph-dokan uses a hardcoded dokan timeout of 120 seconds. We'll
make it configurable and set it only once using dokan options,
instead of setting it for every request.

While at it, we're increasing the default timeout to 5 minutes
to be on the safe side when hitting high load, especially since
timeouts are very disruptive as Dokan will unmount the filesytem
on timeout.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>