squashme: use ceph logging and log error codes
We'll have to use ceph logging facilities instead of logging directly
to stderr, otherwise we won't be able to debug daemons running in the
background.

Worth mentioning that the defaults for the log_to_stderr / err_to_stderr
options change based on the service type.

Some messages (e.g. help text, messages that won't be logged by background
processes) can be logged directly to the console. The stadard rbd-nbd
file also does it, so we may want to simplify merging the two
implementations, keeping the code unchanged whenever possible.