rbd: add WNBD utils
The WNBD[1] driver implements the NBD protocol. This DeviceIoControl
interface used for interacting with it is slightly different than
the ioctl Linux NBD interface.

This change will allow opening, listing, mapping and unmapping
WNBD devices.

Note that while on Linux a unix socket handle is passed to the driver,
we cannot do that on Windows. Instead, we're opening a TCP listener
and asking the driver to connect to it.

In the future, we may change the userspace <-> driver communication
channel.

[1] https://github.com/cloudbase/wnbd

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@cloudbasesolutions.com>