rbd: fix import image path parsing on Windows
When importing an image, the rbd command uses only the file name
and expects "/" to be used as a separator. On Windows, it will
use the entire path as image name since the path separator is not
the same.

This change updates it so that the "\\" path separator can be
properly handled as well.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>