rbd-nbd: accept FUA flag
This change updates the RBD NBD server, propagating the Force
Unit Flag (FUA) flag for write requests, ensuring that they are
flushed.

While the SCSI specs allow FUA to be used for read requests as well
in order to ensure that read operations don't use stale data,
NBD currently prohibits it.

In order to avoid unnecessary operations, we'll accept FUA and flushes
only when RBD writeback caching is enabled, advertising it via NBD
flags.

For now, we're only enabling FUA on Windows. If needed, the same can
be done for Linux.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>