wip: wvbd bindings
Using NBD wasn't efficient enough, so we're now using a different
IO transport. In fact, most of the interaction with the driver
is now handled by wvbd.dll.

This change adds a new tool called rbd-wvbd that will handle the
rbd mappings. The Windows rbd-nbd tool will be removed in favor
of this.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>