wip: reject new mappings when stopping
This change disables the "NewMappingsAllowed" WNBD setting when
the Ceph service stops.

This helps ensuring that all WNBD mappings are cleanly disconnected
when the Ceph service stops. The open question is whether Ceph should
actually do this, considering that it's a global setting. It might
also be confusing for users that won't be able to create new mappings
when the ceph service is stopped.