rbd: allow non-persistent Windows mappings
At the moment, all Windows RBD mappings are persistent, being
recreated when the Ceph service starts.

This change adds the "--non-persistent" flag to allow skipping
certain images.

Note that even for non-persistent mappings, we're still storing
data in the Windows registry, allowing us to retrieve image
details such as the rbd pool as well as the admin socket path.

When the daemons stop or the Ceph service starts, non-persistent
entries are cleaned up from the Windows registry.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>