rbd: filter inheritted child proc handles on Windows
When spawning rbd-wnbd daemons, we're using anonymous pipes to
report back the status of the operation. The issue is that we're
spawning multiple processes at a time and the pipe handles are
inherited by multiple processes.

If a process returns without writing back, reading from the pipe can
hang as there still are other processes keeping the handle open.

This patch will filter the handles that are inheritted by each
child process when starting rbd-wnbd daemons.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>