cephfs: add ceph_may_delete function
We're adding a new libcephfs function: ceph_may_delete. It checks
if the mount permissions allow deleting a file or directory, without
actually deleting it.

This will allow us to drop the redundant permission checks at
ceph-dokan level, saving about 1500 LOC.

Note that the "DeleteFile" Dokan callback expects us to say if a
delete operation is allowed. The "Cleanup" callback is supposed
to perform the actual file or directory deletion.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>