denc: support non-c++20 compliant C++ standard library
when compiling with the standard library comes with GCC-10, we have
FTBFS like:

```
src/include/denc.h:517:49: error: 'bit_cast' is not a member of 'std';
did you mean 'bad_cast'?
  517 |   unsigned lowznib = v ?
      (std::countr_zero(std::bit_cast<uint64_t>(v)) / 4) : 0u;
      |                                                 ^~~~~~~~
      |                                                 bad_cast
```

to address this issue, an implementation of std::bit_cast<> is defined
if it is not available. in the long run, we should use a better C++
compiler for compiling the tree.

Fixes: https://tracker.ceph.com/issues/57355
Signed-off-by: Kefu Chai <tchaikov@gmail.com>