win32_build.sh: build "ceph_test*"
At the moment, win32_build.sh builds the "tests" target, which
only includes the "unittest_*" tests.

We'll update the build script to include "ceph_test_*" targets
as well.

Note that we can't just build all the targets that were generated
by cmake since some of them are unavailable on Windows. When
doing the Windows port, we decided to avoid polluting the Cmake
files and keep the amount of Windows checks to a minimum.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>