wip, rbd_wnbd: persistent reservation support
We're updating rbd-wnbd to accept scsi persistent reservation
commands, which are going to be stored using rados extended
attributes.

The workflow resembles the target_core_rbd linux kernel module,
with a few differences:

* we're storing the PR data using ceph encoding
* the hostname is used as i_t nexus identifier

This is a work in progress, we'll need to properly implement
the scsi commands. Also, we'll have to check if a locking/fencing
mechanism is required.

Limitations:

* the hostname is used as i_t nexus identifier
* mainly aimed at MS Failover Cluster / CSVs, maay not be useful
  for scenarios that involve containers or VM passthrough disks
* only Windows hosts that use rbd-wnbd will be aware of the
  persistent reservations
* only SPC-3 is supported (also pinned by the WNBD driver)
* other initiators aren't notified through SENSE status when
  the reservations are released / preempted