rbd: fix rbd-wnbd const warning
libwnbd receives a structure containing callbacks for various IO
operations. We're declaring it as a static member of the
WnbdHandler class:

  static constexpr WNBD_INTERFACE RbdWnbdInterface

WnbdCreate receives a const PWNBD_INTERFACE argument and mingw
throws a warning about the const qualifier:

  wnbd_handler.cc:419:33: warning: invalid conversion from
  ‘const WNBD_INTERFACE*’ {aka ‘const _WNBD_INTERFACE*’} to
  ‘PWNBD_INTERFACE’ {aka ‘_WNBD_INTERFACE*’} [-fpermissive]

We'll avoid the warning by adding an explicit const cast, which is
safe considering the RbdWnbdInterface declaration.