librbd: add optional "assert_tag" parameter to write ops
In order to emulate SCSI persistent reservations, we need to be able
to invalidate IO operations after a reservation gets preempted.

We'll add an optional "assert_tag" parameter that can be used to
ensure that the image tag (persistent reservation version) didn't
change.

In case of a tag mismatch, the IO requestor is supposed to refresh
the persistent reservation data and either return a reservation
conflict error or retry the operation, depending on the current
reservations.

TODO: do the same for "read" requests.