rbd-wnbd: allow waiting for disks to become available
After a WNBD mapping is created, the driver informs Storport that
the bus changed, expecting it to rescan the bus and expose the disk.
This is an asynchronous process and it usually takes a matter of
milliseconds, up to a few seconds under significant load.

The fsx librbd test ocasionally fails to open the disk as it isn't
ready yet. Unlike the Python rbd-wnbd test, it doesn't perform
any polling.

For convenience, WNBD now includes a helper function called
WnbdPollDiskNumber. We'll add some rbd-wnbd map parameters
(--wait-online and --wait-online-timeout) so that rbd-wnbd callers
won't have to do the polling theselves.

We'll update the fsx test to use those parameters, but not the
Python test, which dumps useful profiling information.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>